feat(update): first version of update hook
To use the hook on the server, we must use an update hook instead of an
commit-message hook. This commit provides a version version of an
update hook. To extract the commit messages from an update hook
script, we need access to the git binary and parse some git-log outputs